V1.01

+++ PR3-Blazor/Components/Models/Poste.cs
@@ -2,10 +2,10 @@
 {
     public class Poste
     {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Numero { get; set; }
     public string MacAdress { get; set; }
-    public long SalleId { get; set; }
+    public int SalleId { get; set; }
     public bool IsConnected { get; set; }
 
     }
